added created by in post, category checking when deleting

+++ backend/app/Http/Controllers/CategoryController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use Illuminate\Http\Request;
+use App\Services\PostService;
+use App\Utilities\StatusCodes;
 use App\Services\CategoryService;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Auth;
@@ -12,9 +14,11 @@ use App\Http\Requests\Category\CategoryCreate;
 class CategoryController extends Controller
 {
     private $service;
+    private $postService;
 
-    function __construct(CategoryService $service) {
+    function __construct(CategoryService $service, PostService $postService) {
         $this->service = $service;
+        $this->postService = $postService;
     }
 
     public function create(CategoryCreate $request) {
@@ -43,6 +47,13 @@ class CategoryController extends Controller
     public function destroy(Request $request, $id) {
         $category = $this->setCategory($id);
         $this->authorize("destroy", $category);
+
+        $used = $this->postService->findByCtgry($category->id);
+
+        if (count($used) > 0) {
+            return $this->response((object)[], "Category cannot be deleted, it is being used by your post", StatusCodes::MIDDLEWARE);
+        }
+
         $result = $this->service->destroy($id);
 
         return $this->response($result, "Category has been deleted Successfully...!");
